+++ public/javascripts/comment-edit.js
@@ -2,7 +2,7 @@
 //
 //Referenced by comment.scala.html
 //
-function editComment(commentId, commentText, reloadPage){
+function editComment(commentId, commentText, senderName, senderEmail, commentLink, reloadPage){
 	
 	if(reloadPage === undefined) reloadPage = false;
 	
@@ -17,6 +17,18 @@ function editComment(commentId, commentText, reloadPage){
     
     request.done(function (response, textStatus, jqXHR){
         console.log("Response " + response);
+        $('#editField_' + commentId).mentionsInput('getMentions', function(data) {
+            // Send email to any users tagged in this comment, and subscribe them to this resource
+            data.forEach(function(mentioned){
+                var text = senderName+' mentioned you in a comment: '+commentLink+'\n\n';
+                text += commentText;
+                request = jsRoutes.controllers.Users.sendEmail("New Clowder comment", senderEmail, mentioned.email, text).ajax({
+                    type: 'POST'
+                }).done(function(response){
+                    console.log(response);
+                })
+            })
+        });
         //Sucessful update of the DB - update the interface
     	$("#comment-body_" + commentId).html(theText.replace(/\n/g, "<br>"));
     });